requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Werkzeug==2.3.7
SQLAlchemy==2.0.21
WTForms==3.0.1
email-validator==2.0.0
PyJWT==2.8.0
bcrypt==4.0.1
cryptography==41.0.4

# Production dependencies
gunicorn==21.2.0
psycopg2-binary==2.9.7
redis==5.0.0
celery==5.3.1

# AI/ML dependencies
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1
plotly==5.15.0
folium==0.14.0

# Data visualization
matplotlib==3.7.2
seaborn==0.12.2

# API and web services
requests==2.31.0
urllib3==2.0.4

# Security
python-dotenv==1.0.0
cryptography==41.0.4

# Monitoring and logging
sentry-sdk[flask]==1.28.1
structlog==23.1.0

# Testing
pytest==7.4.0
pytest-flask==1.2.0

# Development
black==23.7.0
flake8==6.0.0 